Conflicts:
	lib/graph_commit.rb

+++ app/controllers/admin/users_controller.rb
@@ -1,61 +1,53 @@
-class Admin::UsersController < ApplicationController
-  layout "admin"
-  before_filter :authenticate_user!
-  before_filter :authenticate_admin!
+class Admin::UsersController < Admin::ApplicationController
+  before_filter :admin_user, only: [:show, :edit, :update, :destroy]
 
   def index
     @admin_users = User.scoped
     @admin_users = @admin_users.filter(params[:filter])
-    @admin_users = @admin_users.order("updated_at DESC").page(params[:page])
+    @admin_users = @admin_users.search(params[:name]) if params[:name].present?
+    @admin_users = @admin_users.alphabetically.page(params[:page])
   end
 
   def show
-    @admin_user = User.find(params[:id])
+    # Projects user can be added to
+    @not_in_projects = Project.scoped
+    @not_in_projects = @not_in_projects.without_user(admin_user) if admin_user.authorized_projects.present?
 
-    @projects = if @admin_user.projects.empty?
-               Project
-             else
-               Project.without_user(@admin_user)
-             end.all
+    # Projects he already own or joined
+    @projects = admin_user.authorized_projects.where('projects.id in (?)', admin_user.authorized_projects.map(&:id))
   end
 
   def team_update
-    @admin_user = User.find(params[:id])
-
-    UsersProject.user_bulk_import(
-      @admin_user, 
+    UsersProject.add_users_into_projects(
       params[:project_ids],
+      [admin_user.id],
       params[:project_access]
     )
 
-    redirect_to [:admin, @admin_user], notice: 'Teams were successfully updated.'
+    redirect_to [:admin, admin_user], notice: 'Teams were successfully updated.'
   end
 
 
   def new
-    @admin_user = User.new(:projects_limit => 10)
+    @admin_user = User.new({ projects_limit: Gitlab.config.gitlab.default_projects_limit }, as: :admin)
   end
 
   def edit
-    @admin_user = User.find(params[:id])
+    admin_user
   end
 
-  def block 
-    @admin_user = User.find(params[:id])
-
-    if @admin_user.block
+  def block
+    if admin_user.block
       redirect_to :back, alert: "Successfully blocked"
-    else 
+    else
       redirect_to :back, alert: "Error occured. User was not blocked"
     end
   end
 
-  def unblock 
-    @admin_user = User.find(params[:id])
-
-    if @admin_user.update_attribute(:blocked, false)
+  def unblock
+    if admin_user.update_attribute(:blocked, false)
       redirect_to :back, alert: "Successfully unblocked"
-    else 
+    else
       redirect_to :back, alert: "Error occured. User was not unblocked"
     end
   end
@@ -63,7 +55,7 @@ class Admin::UsersController < ApplicationController
   def create
     admin = params[:user].delete("admin")
 
-    @admin_user = User.new(params[:user])
+    @admin_user = User.new(params[:user], as: :admin)
     @admin_user.admin = (admin && admin.to_i > 0)
 
     respond_to do |format|
@@ -85,27 +77,34 @@ class Admin::UsersController < ApplicationController
       params[:user].delete(:password_confirmation)
     end
 
-    @admin_user = User.find(params[:id])
-    @admin_user.admin = (admin && admin.to_i > 0)
+    admin_user.admin = (admin && admin.to_i > 0)
 
     respond_to do |format|
-      if @admin_user.update_attributes(params[:user])
-        format.html { redirect_to [:admin, @admin_user], notice: 'User was successfully updated.' }
+      if admin_user.update_attributes(params[:user], as: :admin)
+        format.html { redirect_to [:admin, admin_user], notice: 'User was successfully updated.' }
         format.json { head :ok }
       else
         format.html { render action: "edit" }
-        format.json { render json: @admin_user.errors, status: :unprocessable_entity }
+        format.json { render json: admin_user.errors, status: :unprocessable_entity }
       end
     end
   end
 
   def destroy
-    @admin_user = User.find(params[:id])
-    @admin_user.destroy
+    if admin_user.personal_projects.count > 0
+      redirect_to admin_users_path, alert: "User is a project owner and can't be removed." and return
+    end
+    admin_user.destroy
 
     respond_to do |format|
-      format.html { redirect_to admin_users_url }
+      format.html { redirect_to admin_users_path }
       format.json { head :ok }
     end
+  end
+
+  protected
+
+  def admin_user
+    @admin_user ||= User.find_by_username!(params[:id])
   end
 end
